feat(api): add middlewares and restructure everything
