+++ src/utils/functions.ts
@@ -1,3 +1,4 @@
+// import { findCookiesByCookies } from '@/api/models/cookies'
 import { Feedback, FeedbackRef } from '@/types'
 import { App } from '@capacitor/app'
 import { computed } from 'vue'
@@ -75,11 +76,11 @@ export function createSeasons(): string[] {
 
 /**
  * Handle the hardware back button
- * 
+ *
  * @param condition If true then callback, if false then window.history.back()
  * @param callback What to do when button is pressed
  */
-export function handleBackButton(condition: ()  => boolean, callback: () => void): void {
+export function handleBackButton(condition: () => boolean, callback: () => void): void {
   App.addListener('backButton', () => {
     if (condition()) {
       callback()
